remove example
